chore: typer 阶段清零（capabilities/网络/client/server/integration/util/data 各子树完成），进入 refchecks 阶段

+++ src/main/scala/li/cil/oc/common/component/traits/TextBufferProxy.scala
@@ -156,7 +156,7 @@ trait TextBufferProxy extends api.internal.TextBuffer {
       for (x <- col until ((col + line.length) min data.width)) {
         val packedBackground = data.color(y)(x) & 0x00FF
         val packedForeground = (data.format.deflate(PackedColor.Color(line(x - col))) << PackedColor.ForegroundShift) & 0xFF00
-        data.color(y)(x) = packedForeground | packedBackground
+        data.color(y)(x) = (packedForeground | packedBackground).toShort
       }
     }
   }
@@ -167,7 +167,7 @@ trait TextBufferProxy extends api.internal.TextBuffer {
       for (x <- col until ((col + line.length) min data.width)) {
         val packedBackground = data.format.deflate(PackedColor.Color(line(x - col))) & 0x00FF
         val packedForeground = data.color(y)(x) & 0xFF00
-        data.color(y)(x) = packedForeground | packedBackground
+        data.color(y)(x) = (packedForeground | packedBackground).toShort
       }
     }
   }
@@ -175,7 +175,7 @@ trait TextBufferProxy extends api.internal.TextBuffer {
   private def color(column: Int, row: Int): Short = {
     if (column < 0 || column >= getWidth || row < 0 || row >= getHeight)
       throw new IndexOutOfBoundsException()
-    else data.color(row)(column).toShort
+    else data.color(row)(column)
   }
 }
